ref(ui): Move Security & Privacy Fields to another file
