+++ src/sentry/static/sentry/app/views/settings/project/navigationConfiguration.tsx
@@ -75,7 +75,7 @@ export default function getConfiguration({
           path: `${pathPrefix}/security-and-privacy/`,
           title: t('Security & Privacy'),
           description: t(
-            'Configuration related to dealing with sensitive data and other security settings. (Data Scrubbing, Data Privacy, Data Scrubbing)'
+            'Configuration related to dealing with sensitive data and other security settings. (Data Scrubbing, Data Privacy, Data Scrubbing) for a project'
           ),
         },
